Added Table to DB for Counters and added methods in MainActivity and CalendarAdapter to use Counters
Using them should be in place.

Need interface for the user to add a counter.
User should be able to input name and info about the counter.

Then each time the app is loaded we need to add "new" counter if the user hasn't stopped the counter.

And we need to have interface for the user to stop / reset a counter when they click on the counter textview on grid and also within the day view.

Additionally we may want to display the counter on the day view.